Added authentication and password reset

+++ routers/posts.py
@@ -1,26 +1,47 @@
 from typing import Annotated
 
-from fastapi import APIRouter, Depends, HTTPException, status
-from sqlalchemy import select
+from fastapi import APIRouter, Depends, HTTPException, status, Query
+from sqlalchemy import select, func
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
 
+from fastapi import Query
+
 import models
 from database import get_db
-from schemas import PostCreate, PostResponse, PostUpdate
+from schemas import PostCreate, PostResponse, PostUpdate, PaginatedPostResponse
+from auth import CurrentUser
 
 router = APIRouter()
 
+# ge - greater than , le - less than = 0,10 default
+@router.get("", response_model=PaginatedPostResponse)
+async def get_posts(
+    db: Annotated[AsyncSession, Depends(get_db)],
+    skip:Annotated[int , Query(ge=0)]=0,
+    limit:Annotated[int, Query(ge=1, le=100)] = 10,
+    ):
 
-@router.get("", response_model=list[PostResponse])
-async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
+    count_result = await db.execute(select(func.count()).select_from(models.Post))
+    total = count_result.scalar() or 0 
     result = await db.execute(
         select(models.Post)
         .options(selectinload(models.Post.author))
-        .order_by(models.Post.date_posted.desc()),
+        .order_by(models.Post.date_posted.desc())
+        .offset(skip)
+        .limit(limit) ,
     )
     posts = result.scalars().all()
-    return posts
+
+    has_more = skip + len(posts) < total 
+
+    return PaginatedPostResponse(
+        posts = [PostResponse.model_validate(post) for post in posts],
+        total = total,
+        skip=skip,
+        limit= limit,
+        has_more=has_more
+    )
 
 
 @router.post(
@@ -28,21 +49,21 @@ async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
     response_model=PostResponse,
     status_code=status.HTTP_201_CREATED,
 )
-async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(
-        select(models.User).where(models.User.id == post.user_id),
-    )
-    user = result.scalars().first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
+async def create_post(post: PostCreate,current_user:CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
+    # result = await db.execute(
+    #     select(models.User).where(models.User.id == post.user_id),
+    # )
+    # user = result.scalars().first()
+    # if not user:
+    #     raise HTTPException(
+    #         status_code=status.HTTP_404_NOT_FOUND,
+    #         detail="User not found",
+    #     ) no need cuz of CurrentUser
 
     new_post = models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id,
+        user_id=current_user.id,
     )
     db.add(new_post)
     await db.commit()
@@ -67,6 +88,7 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
 async def update_post_full(
     post_id: int,
     post_data: PostCreate,
+    current_user: CurrentUser,
     db: Annotated[AsyncSession, Depends(get_db)],
 ):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
@@ -76,20 +98,26 @@ async def update_post_full(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Post not found",
         )
-    if post_data.user_id != post.user_id:
-        result = await db.execute(
-            select(models.User).where(models.User.id == post_data.user_id),
-        )
-        user = result.scalars().first()
-        if not user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="User not found",
-            )
+    # if post_data.user_id != post.user_id:
+    #     result = await db.execute(
+    #         select(models.User).where(models.User.id == post_data.user_id),
+    #     )
+    #     user = result.scalars().first()
+    #     if not user:
+    #         raise HTTPException(
+    #             status_code=status.HTTP_404_NOT_FOUND,
+    #             detail="User not found",
+    #         ) cuz of current_user
+
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="not authorised to update this post",
+        )   
 
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
+    # post.user_id = post_data.user_id
 
     await db.commit()
     await db.refresh(post, attribute_names=["author"])
@@ -100,6 +128,7 @@ async def update_post_full(
 async def update_post_partial(
     post_id: int,
     post_data: PostUpdate,
+    current_user: CurrentUser,
     db: Annotated[AsyncSession, Depends(get_db)],
 ):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
@@ -109,6 +138,13 @@ async def update_post_partial(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Post not found",
         )
+
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="not authorised to update this post",
+        )   
+
 
     update_data = post_data.model_dump(exclude_unset=True)
     for field, value in update_data.items():
@@ -120,7 +156,7 @@ async def update_post_partial(
 
 
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(post_id: int, current_user:CurrentUser,db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
@@ -128,6 +164,12 @@ async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Post not found",
         )
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="not authorised to delete this post",
+        )   
+
 
     await db.delete(post)
     await db.commit()
